Move await/spawn handlers into doeff.handlers package
